split core messaging and bsb
