+++ components/bsb/timer/main/impl/build.gradle
@@ -30,7 +30,7 @@ commonDependencies {
     implementation(projects.components.bsb.timer.cards.api)
     implementation(projects.components.bsb.timer.done.api)
 
-    implementation(projects.components.bsb.wear.messenger)
+    implementation(projects.components.bsb.wear.bridgeBsb)
 
     implementation(libs.decompose)
     implementation(libs.decompose.composeExtension)
