+++ components/bsb/wear/bridge-bsb/build.gradle
@@ -1,0 +1,21 @@
+plugins {
+    id("flipper.multiplatform")
+    id("flipper.anvil-multiplatform")
+    id("flipper.multiplatform-dependencies")
+    id("kotlinx-serialization")
+}
+
+androidDependencies {
+    implementation(projects.components.core.di)
+    implementation(projects.components.core.ktx)
+    implementation(projects.components.core.data)
+    implementation(projects.components.bsb.preference.api)
+    implementation(projects.components.bsb.timer.background.api)
+    api(projects.components.bsb.wear.bridge)
+
+    implementation(libs.kotlin.serialization.json)
+    implementation(libs.google.horologist.datalayer)
+    implementation(libs.google.horologist.datalayer.watch)
+    implementation(libs.google.horologist.datalayer.phone)
+    implementation(libs.kotlin.coroutines.gplayservices)
+}
