+++ components/bsb/wear/bridge-bsb/src/androidMain/AndroidManifest.xml
@@ -3,7 +3,7 @@
 
     <application>
         <service
-                android:name=".service.StatusListenerService"
+                android:name="com.flipperdevices.bsb.wear.messenger.service.StatusListenerService"
                 android:exported="true">
             <intent-filter>
                 <action android:name="com.google.android.gms.wearable.MESSAGE_RECEIVED" />
