+++ components/bsb/wear/bridge-bsb/src/androidMain/kotlin/com/flipperdevices/bsb/wear/messenger/model/WearMessage.kt
