+++ components/bsb/wear/bridge/src/androidMain/kotlin/com/flipperdevices/bsb/wear/messenger/application/WearMessengerApplication.kt
