+++ components/bsb/wear/bridge/src/androidMain/kotlin/com/flipperdevices/bsb/wear/messenger/consumer/WearMessageConsumer.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.bsb.wear.messenger.consumer
+
+import com.flipperdevices.bsb.wear.messenger.serializer.DecodedWearMessage
+import com.flipperdevices.bsb.wear.messenger.serializer.WearMessageSerializer
+import kotlinx.coroutines.flow.Flow
+
+interface WearMessageConsumer {
+    val messagesFlow: Flow<DecodedWearMessage<*>>
+    fun <T> consume(message: WearMessageSerializer<T>, byteArray: ByteArray)
+}
