+++ components/bsb/wear/bridge/src/androidMain/kotlin/com/flipperdevices/bsb/wear/messenger/producer/WearMessageProducer.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.bsb.wear.messenger.producer
+
+import com.flipperdevices.bsb.wear.messenger.serializer.WearMessageSerializer
+
+interface WearMessageProducer {
+    suspend fun <T> produce(message: WearMessageSerializer<T>, value: T)
+}
